Work on UI scripting: tabbox fixes and completion
The QTabWidget now has a wrapper for currentChanged, in the form of the OnSelectionChanged event.

Fixed how tabbox tabs are handled; they now subclass widget instead of being a pseudo-wrapper for them. The previous approach failed because we can only have one wrapper per widget, so (tab.body) was returning a tabbox_tab instead of "casting itself" to widget. Since we now subclass widget, we've also prefixed the tab-specific field names.