Work on scripting: form API fixes; pause script while confirming close
Fixed several errors in the Lua APIs for working with forms, based on type checks that became incorrect when we changed how form types are exposed to scripts.

If you try to close the hardcoded script window while a script is running, a confirmation prompt will appear; this isn't new, but what's new is that script execution is now paused while the prompt is open.