Prototyping for CanvasWidget
Did some prototyping for creating a CanvasWidget meant to render layers of arbitrary data; currently it only supports QImages, but in the future, we'd like to support text and maybe SVG. This is meant to be similar to Qt's "graphics scenes," but with the ability to have all used resources be managed entirely externally (e.g. by our script VM).