R&D: Investigation into a planned resource manager for Lua
Wrote some tests, with the goal of determining whether it would be possible to have Lua centrally manage script-loaded raster resources that could then be shared by the UI. Test results indicate that we can easily manage resource lifetime and ensure resource sharing even by model items, with virtually no overhead; however, ensuring that widgets live-update as resources are modified is harder, and is likely to incur a significant burden for large models that use lots of resources (e.g. an entire column that has icons, in a table with many rows).