Work on UI scripting: started work on the resource manager
Started work on a script VM subsystem for managing raster resources. The aim of this system is to allow scripts to use rasters in multiple places, including from within widgets and model items, with the rasters being deleted only when they are unused.

We currently have the raw basics of the system implemented, but not compiled or tested. There are some changes we need to make to other parts of the Lua VM, e.g. to let Lua wrappers actually refer to these resources. Then, we need to make an API for loading game raster assets into rasters, and add an API for setting such rasters as the icons on dropdown items (that'll just be the fastest way to get them into the UI at all, for testing).