Work on UI scripting: dropdown item icons
The VM core's main thread handler now calls the resource subsystem's main thread handler.

When the resource subsystem updates widgets, it now uses special handling for QComboBoxes.

Added a convenience function to LuaManagedResourceHandle for extracting a LuaManagedResource pointer from a QVariant which wraps a resource handle.

Dropdown items can now be given rasters to serve as icons.

Added a DovahKitScriptItemDelegate class, to be used by scripted dropdowns and possibly other widgets. Currently, it allows LuaManagedResourceHandles in the Qt::DecorationRole to display as icons.

When creating rasters, you can now supply the background color to fill with.