Work on UI scripting: work on Lua-managed resources
Fixed initialization of several Lua classes including all widget classes. We were redundantly defining the class metatable.

Modified our class helpers to guard against the above mistake by asserting that the class-to-be-defined does not already exist.

We no longer use luaL_ref and luaL_unref for our wrapper storage.

The `raster` class in Lua now properly inherits our `!base` class and its `__gc` metamethod. That said, tests indicate that for rasters only, Lua garbage collection is not occurring, and I have been unable to figure out why.

Added code to the VM core to clear the resource subsystem's resources when tearing down the VM. Unfortunately, we can't seem to properly control the order in which widgets and model items are torn down, so this is resulting in LuaManagedResourceHandles trying to act on already-deleted resources.