Work on UI scripting: work on Lua-managed resources
Failed attempt at fixing use-after-free issues with LuaManagedResourceHandle, stemming from our apparent inability to control the order in which VM-relevant QObjects are deleted (i.e. we cannot ensure that all widgets and model items which are using a LuaManagedResource are deleted before the LMR itself is deleted, and if deletion occurs out of order, then LMRHs extant in the widgets use the LMR after it has been freed).

Hooking the QObject::destroyed signal, clearly, is not reliable in this particular situation, though I have no idea why. The call stack shows Qt code only -- we're somewhere in event dispatching and rendering -- and nothing in the call stack is easily identifiable as an LMR(H).