Work on UI scripting: work on resource manager simplifications
Currently, when the content of a Lua-managed resource is modified, we track down every widget using it and force those widgets, individually, to repaint; we do this by tracking what models are using what resources. This tracking is expensive and adds significant complexity to the VM, so I'm investigating the possibility of simply forcing all scripted windows and their descendants to repaint.

Calling QWidget::update on a window will force its contained children to repaint; however, this behavior isn't described in Qt's documentation, and writings around the web suggest that it could be OS-dependent. Moreover, a combobox's "tray" (i.e. its view) doesn't repaint when the combobox itself does, because it's a separate widget and isn't part of the normal widget hierarchy; it is not, strictly speaking, a child of the combobox.

It's possible for an event filter to forward paint and update events from a combobox to its view; however, this approach feels messy, and would affect all repaints and not just ones specific to what we're trying to do; further, it relies on updates descending down through widget hierarchies which, again, doesn't seem to be guaranteed behavior.

As such, I've created a helper function which will simply find all widgets in a hierarchy (starting at some root) and fire QWidget::update individually on each of them, additionally checking whether any of them are QComboBoxes and if so, manually queuing updates on their views as well. With this commit, I believe we can next remove all of the model-tracking stuff from the scripted resource subsystem.