Work on UI scripting
Added an MOPH for table cell icons that allows you to specify a raster or any supported color format to serve as the cell's icon.

These tests establish that our method of forcing widgets to repaint should work on dropdowns and table views. We can proceed in removing the bookkeeping associated with the old method.