BSA reading fixes; work on resource scripting
Started work on a dovah.lookup_game_asset function which retrieves any game asset as a Lua-accessible resource. Unrecognized resources are stored as simple QByteArrays; in theory we could offer raw binary access to them in Lua, perhaps mimicking JavaScript's DataView.

Removed a redundant loose file lookup in DovahKitCore::lookup_game_asset.

Fixed the process of reading a file out of a BSA: when filenames are retained, we were skipping the length of the filename, but not the length of the "length" byte itself.