Work on DDS support: added DirectXTex
Added DirectXTex to the program.

DovahKitCore now sets up COM on the main thread, so that DirectXTex can function.

The Lua API to look up game assets now detects DDS files by extension and uses DirectXTex to convert them. Though we don't encounter any apparent errors, testing indicates that we aren't getting valid image data: the QImages are not null, but when used as LMRs in a dropdown, nothing appears.