Work on UI scripting: work on DDS scripting
LuaManagedResources now have an explicit type enum.

You can now wrap a DDS file in a LuaManagedResource and query its basic properties.

Added a Lua class for DDS LMRs. The class is barebones and offers no members, so it's not yet possible to extract and work with raster data within script.

The styled item delegate that we use for LMRs now supports DDS LMRs.

The resource subsystem is now notified whenever an LMR becomes (un)referenced within Qt. In theory, we could use this as a means to only resynchronize resources that are actively being displayed in the UI somewhere.

The MOPH functions for icons now accept DDS LMRs.

The function to look up a game asset will now return a DDS LMR if it finds a valid DDS file.