Work on UI scripting: added image_widget; size [gs]etters
Scripts can now set a widget's minimum and maximum sizes, with the values testing as nil if no (practical) limit exists.

Fixed a bad assertion present when pushing DDS resources into Lua.

Made the LuaManagedRasterWidget available to Lua as `image_widget`.

Fixed the LuaManagedRasterWidget so that it will maintain the raster's aspect ratio when rendering.