BSA fixes
When we load BSA files, we now check our reads to ensure that we're not reading past the end of the file. We also preemptively flag as "corrupt" any file that claims to extend past the end of the file.

BSA file loading now relies on exceptions instead of an error enum. Of course, our threaded loader just catches and discards these. Eh.

The BSA-reading code now checks to ensure that its cobb::mapped_file actually opened properly.

Fixed how we read the BSA lists out of Skyrim.ini, as leading spaces on the last file in a list would be retained within the filename, causing a missing file error.