Work on UI scripting
When a Lua wrapper is garbage-collected, it will now contact the userdata subsystem and destroy its containing list if that list is empty (i.e. if that was the last wrapper for that pertinent pointer). This will allow us to avoid having an empty wrapper list keyed to what may be, at that point, a dangling pointer. Moreover, we can potentially co-opt this to avoid having to maintain VM-side Lua refcounts for referenced objects (as we currently do for model observers), as we'll know for certain that the Lua refcount is zero when we find that the wrapped object's wrapper list is empty.

Made the canvas widget available to Lua. No APIs are offered for it yet, so it's not yet usable.