Work on scripting: work on model observer lifetime management
The script VM no longer tracks Lua-side refcounts for model observers. Instead, we rely on the userdata service's `prune_wrapper_list_for` function to tell us when a model observer is no longer Lua-referenced.

We now divide model observers into an "extant" list of Lua-referenced observers, and a "pending deletion" list of Lua-unreferenced observers, with a thread-lock on the latter so that these observers can be safely deleted from the main thread.

In the future, we can and should expand this to cover widgets as well. Currently, widgets always run full lifetime/teardown code whenever any userdata pointing to the widget is destroyed. You would probably expect this to cause issues when multiple userdata refer to [different parts of] the same widget, but it doesn't: the function to check whether the widget's containing widget tree is still referenced will see the remaining userdata; still, running those checks is a waste of time.