Fixed XLOC; internal documentation
Fixed our code for reading ExtraLock for SSE. We failed to account for struct-formIDs being displaced forward due to their need to be 8-byte-aligned. There may be other places where we've messed this up...

Made corrections to our documentation on the form-loading process.

Did a write-up on form versions. This is intended mainly for outside audiences and is not DovahKit-specific, which is one of the reasons it's markdown-formatted.