Removed struct_form_reference_t
The problem that struct_form_reference_t was meant to solve only occurs in specific places where Bethesda forgot to prevent it, such as WEAP/CRDT; as far as we know, all other such structs are read field-by-field in SSE to match the Classic data format. Moreover, struct_form_reference_t as currently implemented would not have actually allowed us to handle those cases properly: it automated the process of reading padding bytes after a form ID, but the form IDs themselves have different data alignments in Skyrim Special since they're union'd with eight-byte pointers, which means that padding *before* the form ID would've needed to be handled as well.

We could in theory automate the above by making it possible for a subrecord to know when it's in a struct (and how far into a struct it is), but it's not worth implementing the machinery for this when we can just account for padding by hand in the exceptionally rare cases (only one known case) where this issue comes up.

Also deleted the form version document; I was writing that for a public audience, and it's now published elsewhere.