Fixes for last commit; work on UI scripting (lifetime management)
Fixes for last commit: there were some references to struct_form_reference_t that I forgot to remove.

Removed the script core's unused _model_observer class. Previously, it had a refcount on it, and this was how we managed Lua lifetimes for model observers.

We now detect when widgets and button groups become unreferenced within Lua the same way we detect when model observers become unreferenced within Lua. This should allow us to avoid redundant uses of the hierarchy finder when multiple userdata refer to the same widget or button group (i.e. when the wrapper "part" system is in use).