Work on UI scripting: work on object lifetime management
Lua-managed resources now use an "is Lua-referenced" bool, which is managed through the same referenced/unreferenced process as everything else. The only reason there's even a bool at all is because the resource subsystem needs to be able to check whether an LMR is unreferenced while on the client thread.

Overhauled the "hierarchy finder" used to determine whether an object and all connected widget hierarchies is Lua-unreferenced. This overhaul should make it significantly easier to make the hierarchy finder capable of handling new kinds of "bridge" objects in the future.