Work on UI scripting: work on canvas scripting
Created the CanvasWidgetLayerDataLuaManagedResource class, a subclass of CanvasWidgetLayerData. We will use this for raster layers in scripts.

Added the (canvas.layers) collection to Lua, as well as a (canvas:append_layer) API.

Added the (canvas_layer.canvas) getter to Lua.

A CanvasWidgetLayerData can now trigger an update on the layers using it.