Work on UI scripting: work on canvas scripting
Added a new wrapper type for CanvasWidgetLayerData (CWLD) objects that scripts are meant to work with directly. Added a generic CWLD subclass for these, LuaScriptableCanvasWidgetLayerData, which has a Lua refcount and will serve as a superclass for any scriptable CWLDs. The class needs to track whether it's Lua-referenced because we hook CWLD::detached, which will fire on the client thread where we can't check the Lua userdata storage.

The script core is now able to store CWLDs.

Added getter/setter pairs for a canvas widget's image size.