Work on CanvasWidget
CanvasWidget now handles layer blend modes correctly, though I'm no longer certain that it handles layer positioning properly. We'll work on that next.

To achieve this, we had to manually implement a workaround to get the "multiply" blend mode to behave as we expected. We also had to rely on QImages instead of QPixmaps, both because QPainter's composition modes work best with a particular pixel format and QPixmap doesn't let us select the pixel format, and because alpha transparency in QPixmaps breaks when you're drawing multiple pixmaps on top of (or into) one another. Don't even ask me why. Just don't.