Work on UI scripting: work on canvas scripting
Fixed the text-layer-data so that it now draws properly: when it's not drawing text into a QRect, it uses the specified Y-position for the text baseline rather than the top edge, so we have to manually compensate for that.

Modified the text-layer-data to properly adjust the text alignment used for drawing whenever text bounds are constrained only on one axis.

With these changes, text is now drawn properly at least when it's not constrained to any bounds (such that word-wrapping doesn't meaningfully apply).