Work on a Lua wrapper for QFont
I copied the MOPH code for use as a base, but because QFonts embedded into Qt can't reliably be modified (often you need to copy the QFont, edit your copy, and then write it back), we would need to split operations much more granularly... and that feels messy.

Went back to experimenting with a template metaprogramming approach. Our current code looks like it could actually be viable.

As part of the metaprogramming above, I've created a "function traits" header.

Looks like I forgot to commit the "Qt ownership" helper header that was used by the canvas in past commits, so that's in here, too.