Work on UI scripting: work on canvas text data
Added a `font` setter to the Lua `canvas_text_data` class, and updated our test script to use it.

When the font class is asked to pull a QFont from the Lua stack, it now validates the value at the specified stack position and throws a Lua error if necessary.

Changed how CanvasWidgetLayerDataLuaText measures the rendered size of its text, to ensure consistency with QPainter and prevent a few issues I saw in testing.