Work on LAND scripting
Created a Lua wrapper for landscapes, and fixed the setup for the cell and worldspace wrappers.

You can now get a cell's grid coordinates within its containing worldspace.

Added a new form stub helper function to get a cell's landscape.

Started work on a Lua script to draw a worldspace's heightmap.

An assertion will no longer fail if we fail to create a raster when Lua asks us to (e.g. because the requested size was absurdly large and QImage internals caught that).