LAND backend and scripting fixes; raster scripting fixes
We were decoding landscape heightmap data incorrectly. That's now been fixed, though unfortunately I'm not yet sure how I would go about enforcing the file format rules for changes to the in-memory data.

We've committed to sorting the data such that positive numbers point northeast and negative, southwest; this is consistent not only with how Bethesda stores the land data, but with data on worldspaces and probably with the game's overall coordinate system.

The raster:set_pixel API now bounds-checks the coordinates given as arguments.

APIs that require color arguments now bounds-check RGB(A) values passed in as table members.