Misc fixes to cell; work on drawing APIs
Added a helper function `protected_pull_color` to allow Lua APIs to pull a QColor out of Lua and catch any errors that occur, so that they can handle, ignore, or report those errors in a manner that befits their specific needs.

Added missing permission checks to some `cell` setters.

Added an untested function, `raster:draw_line`, and added some of the helper functions that we'd need for a `raster:draw_rect` function.