Work on raster_draw_path:arc_to; vector2 fixes
Fixed `vector2:length()` and `vector2:length_squared()`; they were accessing the wrong stack offsets for the X and Y values.

The `vector2:rotate(d)` function now type-checks its argument, and uses the right stack offset for the Y value.

Fixed `raster_draw_path:arc_to` for specific cases. The math we were using allows for two possible circles on either side of our lines, so we needed code to determine which side was the correct (i.e. inner) side. Additionally, I've added a quick hack to the code we use to compute the sweep and start angles, though I'm still not sure about it.

Created a Lua test script dedicated just to testing `arc_to`. The idea is that we can set booleans to quickly flip the drawn images, and we can use a single number to rotate the entire drawn image and test different angles -- so that we can figure out the exact conditions under which we need to negate the start and sweep angles for the Qt arcTo call. Unfortunately, our `vector2:rotate(deg)` implementation seems especially prone to floating-point imprecision, which grows larger when rotating by larger angles, and this is compromising our test.