Added backend support for LTEX; file_writer tweaks
Added backend support for LTEX.

Change that slipped out of a previous commit: if we save a worldspace that has persistent refs but no persistent cell (e.g. resaving a malformed file, or bad behavior by a frontend), we now write the persistent refs into a persistent child GRUP for their parent cells, to ensure that they aren't lost.