Work on landscape scripting
LTEX has been exposed to scripts, though it hasn't been tested.

Landscape vertex color now has Lua [gs]etters, though they're not tested.

Landscape quads are now accessible to Lua, with their default textures accessible. This is not yet tested. The intention is to make blend layers accessible through the quads as well, and to have a function on the landscape object that tells you which quad a given x/ygrid coordinate pair lies in.