Tests for LMR performance
I wanted to see if I could optimize LMR modification performance: when the script thread wants to modify an LMR, we need to lock with the main thread to ensure that an already-desynchronized LMR doesn't try to resynchronize during a second modification. The easiest way to prevent this is to stay locked for the duration of the entire modification operation, but an alternate approach is to lock long enough to remove the LMR from the "desynchronized" list, unlock, modify the LMR, and then lock again to add it to the "desynchronized" list.

In tests with our heightmap script, it seems we're actually three seconds slower using the latter approach.