Fixes to script raster APIs and similar
QImage::scaled silently changes an image's format to ARGB32_Premultiplied. The `raster:resize` and `raster:scale` Lua APIs have been adjusted to compensate for this. This fixes the issue with our heightmap script's output looking too dark.

The desired QImage format for script content is now defined in the LMR header and referenced from elsewhere. I've changed it from ARGB32 to ARGB32_Premultiplied because the latter is supposed to be more efficient for QPainter's composition modes and therefore for CanvasWidget. However, this necessitates qPremultiply and qUnpremultiply calls wrapping access to individual pixels, and testing suggests that this has slowed our heightmap script down. It may be more efficient to use ARGB32 and eat a very minor slowdown within CanvasWidget so that the script can process and render things faster. I should test that.

Added a hook to the heightmap script that I can use to adjust land texture colors for output. It's currently unused, but I'm considering multiplying color components in mimicry of sky/sun lighting. Though the heightmap is now much more accurate than before thanks to our fixes re: premultiplied colors, it's still a bit on the grey and dark side, and I suspect that may just be because we aren't applying anything to mimic the game's own sunlight.