Switched the script image format back to ARGB32
Switched the preferred QImage format for scripted rasters from ARGB32_Premultiplied back to ARGB32.

Premultiplied alpha would technically be faster for rendering, except that when dealing with scripted rasters or LMRs in general, CanvasWidget isn't rendering the QImage available to the script thread. CanvasWidget and friends are rendering the content made available to the client thread, which in practice is a QPixmap, so the format of the source QImage is irrelevant.

Meanwhile, the use of a premultiplied-alpha format for the script-side data means that colors passed to or from script need to be unpremultiplied or premultiplied, respectively, for each individual pixel operations. This means that scripts that perform large numbers of pixel operations (e.g. our heightmap script) suffer greatly from the use of premultiplied formats. Changing the format to a non-premultiplied one results in a significant speed boost to script execution.

In fact, even if the main thread could benefit from the use of a premultiplied-alpha format, the penalty to the script thread (i.e. the thread actually doing the work) is a heavy one. As is, CanvasWidget seems to perform robustly enough without having to rely on premultiplied-alpha images as source data. CanvasWidget uses a premultiplied-alpha format internally (i.e. when rendering, layers and layer groups spawn a premultiplied-alpha QImage and ask their data objects to draw onto that), so it benefits as much as possible from premultiplied alpha even when the source data isn't in a premultiplied-alpha format.