Important fixes to form load, persistent cells; heightmap script work
Apparently, I forgot to modify the load process to actually allow worldspaces to track their persistent cells; they had a field for it on the `form_stub_addenda`, but nothing ever set that field. This is now fixed.

Fixed `form_stub::test_record_flags_for_file` to support negative file indices (i.e. -1 is the last file in the list).

The `file_load_order::accept_form_stub` function now properly uses `form_stub::_set_parent_form_one_way` instead of `form_stub::set_parent_form` when loading an override stub. We need to only set one-way relationships during the initial stub load, because we'll be making all relationships bidirectional during the use info build process -- and if a bidirectional relationship already existed, it will be duplicated on the recipient.

Fixed the Lua `formpicker.form_types` setter to accept a single form type. It was always meant to, but it hadn't been updated when form types were changed from integers (encoding four-CCs) to userdata.

Took the test script for heightmaps and used it as the basis for a release-suitable script for heightmaps. The new script is much better organized and easy to edit, and allows the user to choose which worldspace will be rendered. This is intended to be a stable platform for further features, including the ability to outline all cells touched by a given file.