Update worldspace heightmap render.lua
Fixed the case where the lowest layer had a custom blend mode and opacity. These are now forced to normal and 1 when the layer is created, in addition to when layers have their visibilities changed. This allows for easier testing and experimentation (e.g. reordering layers to see how everything looks).