Added miniz; scripted canvas context menu; LMR logging
Added miniz to the project.

Implemented CanvasWidget::drawTo, which can be used to draw a flattened image to an arbitrary paint device via QPainter.

You can now right-click on any scripted canvas to save either a flattened image, or a ZIP archive containing all layers and a plaintext dump of layer and layer group properties. The code for the latter is currently very messy.

Lua scripts can now give names to layers and layer groups.

Debug logging for Lua-managed resources can now be toggled on and off with a constexpr bool or with the _DEBUG macro. This allows for convenient toggling of debugging when testing in Release. Additionally, the deletion of LMRs during teardown is now part of this logging.

Extensive fixes to our heightmap script's code for outlining cells modified in a particular file.