Solution file and miniz; heightmap script work
This commit includes the updated solution file and miniz (including its project), which I forgot to include when I first introduced miniz as part of the canvas export code.

When drawing outlines representing which files edit which cells, the heightmap script now renders lines and fills to separate layers within separate layer groups, which allows us to place all lines above all fills so that they never get blended under a fill. As a bonus, we can now rely on layer opacity for the fill opacity, which saves us the work of having to add an alpha component to an HSL color stored as a string.