Load process now reparents the children of persistent cells
References loaded from a worldspace's persistent cell will now be reparented based on their coordinates during the backend form load process, as is done by the game. This makes it easier for frontend systems to work with worldspaces and their contents. Unfortunately, this change also adds between 1.0 and 1.3 seconds to Skyrim.esm's load time, bringing us from about 3.2 seconds up to the range of 4.2 to 4.5 seconds.

The form_stub::do_custom_parse function now allows lambdas, and the form stub argument passed to the functor is no longer const.

Fixed a race condition in the script VM's task queue: we needed to clear the "empty" flag after locking during `_task_queue::push_back`, not before. Setting it before locking meant that the flag could be cleared in `push_back`, and then cleared in `process` before a task is actually inserted, resulting in the queue telling the outside world that it's empty when it isn't.

The `ui.line.new()` constructor now accepts `"h"` and `"v"` as arguments.

Improvements to the worldspace heightmap script:

The `HeightmapWindow` now remembers all `CellOutlineMap`s produced by the heightmap process, and `CellOutlineIsland`s in turn remember their layers and rasters; these preparations are necessary for allowing users to toggle outline visibility and recolor (by redrawing) outlines through the UI.

For cell outlines, fill opacity is now set on the layer group that contains all fills, instead of individually on each layer.

`CellOutlineIsland`s now require an owner `CellOutlineMap`, from which they pull ancestor layers to append to and the color to draw with.

The script now pre-creates `CellOutlineMap` instances based on the list of files that modify the current worldspace. New `CellOutlineMap`s can be created on demand should we somehow encounter a file that touches a cell but not the worldspace (this should not be possible).

The script now has placeholder code for dealing with persistent references. I noticed a probable case of this in Dawnguard.esm: a single cell on our heightmap seemingly not modified by Dawnguard, surrounded by four that were. My guess was that Dawnguard was placing a persistent reference in that cell.