Fix to form loading re: duplicate form stub source files
Fixed an oversight in the multi-threaded form loading code, which could lead to a threaded loader parsing the record immediately after whatever group it was meant to parse. This issue was resulting in some forms, particularly worldspaces, having duplicate entries in their source file lists.

Modified `form_stub::_add_file` to use a memcpy instead of a loop, when extending its file list; this should hopefully silence an IntelliSense warning (C6386, known for false-positives when dealing with array syntax).