Backend docs; script logging; script window Esc fix
Added three kinds of debug logging to the script core: QObject lifetime logging; model observer lifetime logging; and script start/stop logging. This logging is controlled by constexpr bools, and so can be enabled or disabled in Release (and is auto-enabled in Debug).

Fixed the `font` accessor on scripted QLabels; it was using the wrong wrapper part type, triggering an assertion failure in Debug.

Fixed an oversight that allowed one to close the hardcoded script window out from under a running script, by pressing the Esc key.

Fixed miniz's project settings so that it compiles as a static library in all configurations.

Wrote backend documentation for form stub addenda.