Dovahscript: coordinator Lua setup; userdata
Additional work on the userdata subsystem.

Wrote some of the code for setting up a new Lua state, and some of the code for tearing down a Lua state.

Ported over the code for importing Lua's standard libraries, shimming functions where needed. Ported over the vector3 Lua class as well, as part of reorganizing how Lua setup is performed by the coordinator subsystem.