Dovahscript: lifetime subsystem teardown; event subsystem porting
Wrote teardown code for the lifetime subsystem.

Removed userdata subsystem code for CanvasWidgetEntities. This code was needed as part of special-case code to allow Lua to perform at-will deletion of a layer group (while accounting for its descendants); in the previous engine, CWEs were not allowed to exist as orphans, so this was the only way to allow their removal. In the new engine, we will allow CWEs to be orphaned from CanvasWidgets, and we will have APIs for removing and reparenting them, so they no longer need to be treated as a special case.

Started work on porting over the code for the event subsystem.