Dovahscript: removed the widget metatable override property
Removed "metatable override" functionality from the new script engine, as this functionality can't easily be used from the worker thread: dynamic QObject property accessors are reentrant but not thread-safe, and we can't be sure under what conditions Qt itself might try to get or set properties (though we know they have a defined prefix for Qt-internal properties). If we want widgets of a particular type to use a particular metatable, we'll just have to subclass those widgets.

The override functionality was mainly useful so that horizontal rules could be bare QFrames without every QFrame being locked to the "horizontal/vertical rule" class in Lua. Again, though, it's simpler and easier to subclass them.