Dovahscript: lifetime work for model observers
Created DovahscriptStandardItemModel, a subclass of ObservableStandardItemModel which is intended to be aware of all QWidgets using it (i.e. the script coordinator will make sure it remains informed). This will help with lifetime management, allowing us to treat model observers as bridges akin to QButtonGroups.

Wrote some of the outermost skeletal code needed to perform hierarchy tests on model observers.