Dovahscript: lifetime check code looks complete now
Created the `task_reference_state_multi_checker` struct as something that can be given a passkey and used to "batch" locking and unlocking when checking whether a potentially larger of objects are task-referenced (i.e. as part of a lifetime check).

The lifetime check queue and the hierarchy crawler it spawns now both check whether objects are task-referenced, using instances of the above struct.

The hierarchy crawler can now store model observers as severed bridges. Additionally, it no longer incorrectly uses QObject::deleteLater directly, instead counting on the lifetime system to do that (so that the lifetime system can also stop tracking the to-be-deleted objects and avoid a dangling pointer).

Fixed a mistake in the lifetime system's `destroy_hierarchy_object` member function: we forgot to actually call QObject::deleteLater there.

With the changes in this commit, I believe the essentials of lifetime management are basically done. There are a few minor functions that need to be implemented, e.g. `lifetime::remove_task_reference`.