Dovahscript: host; abort userdata; can actually pass Lua code in now
Created the DovahscriptHost; this is the object that the rest of the program frontend is meant to communicate with when trying to run scripts.

We now use a userdata as our sentinel error when aborting a script. Code for error reporting and the shimmed `pcall` have been modified accordingly.

The resource subsystem now requests a UI repaint through the coordinator subsystem.

There is now a means to pass scripts to the coordinator for execution, and a means to begin a script session. Scripts are now parsed on the worker thread, and multiple files can be executed at a time. The `script_set` struct has an `addon_name` field in preparation for "script packages," a planned feature, but this field should be renamed to `package_folder_name`.