Dovahscript: coordinator porting
The coordinator subsystem now fires several signals directly on the host, and so the host no longer needs to listen for equivalent signals on the coordinator itself.

Renamed `resources::clear` to `resources::on_script_teardown`.

Fixed up the placeholder code in the permissions subsystem.

Ported `coordinator::ui_lock_override`.

Gave `coordinator` a private constructor and destructor.

The coordinator now uses the "abort" userdata error sentinel.

The coordinator now clears its task queues as part of teardown.

Added `coordinator::_internal_scriptDone`, used internally to alert the client thread to the worker thread having finished with all scripts.

Removed all traces of the main-to-script "urgent" queue from the coordinator for now. We really only have one urgent main-to-script message -- that for forms being deleted. If forms are deleted in response to a scripted task, then we need some sort of "reply" mechanism whereby the task can take over and zombify wrappers for that form. If forms are deleted but not in response to a scripted task, then we should just consider that a hard error and call an abort (ideally with an appropriate message).

Wrote internal documentation summarizing the six script engine subsystems.