Dovahscript: simplified coordinator::running
Changed `coordinator::running` to a normal atomic bool, and added a mutex with a specific name to it. Previously, the former was a `cobb::lockable_bool`, a struct which pairs an atomic bool with a mutex, but with a less intuitive locking behavior.

When the old engine is fully deprecated, I can delete `cobb::lockable_bool` altogether. Lookin' forward to it.