Dovahscript: the delete_form task
It is now possible for script-to-client tasks to request control of the script thread for the duration of their execution. This allows these tasks to run code within Lua, e.g. to manage userdata -- an absolute necessity for the `delete_form` task.

The `coordinator` subsystem is now able to assert on unexpected form deletions or form deletion failures.

With these changes, we no longer need an "urgent main-to-script" task queue.

Started work on porting over the `form` class.