Dovahscript: coordinator now has code to set up wrapper classes
We've combined the "define metatable" and "create singleton" steps of the process into a single function, `set_up_all_native_wrappers`.

When we get around to implementing classes that have collections, we can use `wrapper_metatable::has_extra_class_setup` and friends to keep setup code self-contained.

As of this writing, there are no longer any IntelliSense errors in coordinator.cpp. We have to implement some dependencies within our wrappers and APIs, but we're close to being able to test!